requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
redis==5.0.7
django-redis==5.4.0
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.2
xarray==2024.6.0
netCDF4==1.7.1
requests==2.32.3
joblib==1.4.2
scikit-learn==1.5.1
lightgbm==4.5.0
